fix: namespaced StopPoint
