+++ app/graphql/types/tfl/entities/disruption_type.rb
@@ -4,7 +4,7 @@ module Types
       class DisruptionType < Types::BaseObject
         field :additionalInfo, String, null: true
         field :affectedRoutes, [Types::Tfl::Entities::RouteSectionType, null: true], null: true
-        field :affectedStops, [Types::StopPointType, null: true], null: true
+        field :affectedStops, [Types::Tfl::Entities::StopPointType, null: true], null: true
         field :category, Types::CategoryType, null: true
         field :categoryDescription, String, null: true
         field :closureText, String, null: true
